Rename stories, fix misc imports
