+++ client/web/src/enterprise/batches/batch-spec/edit/DownloadSpecModal.story.tsx
@@ -1,10 +1,10 @@
 import { storiesOf } from '@storybook/react'
 
-import { WebStory } from '../../../components/WebStory'
+import { WebStory } from '../../../../components/WebStory'
 
 import { DownloadSpecModal } from './DownloadSpecModal'
 
-const { add } = storiesOf('web/batches/create/DownloadSpecModal', module).addDecorator(story => (
+const { add } = storiesOf('web/batches/batch-spec/edit/DownloadSpecModal', module).addDecorator(story => (
     <div className="p-3 container">{story()}</div>
 ))
 
@@ -14,10 +14,10 @@ add('Download Spec Modal', () => (
             <DownloadSpecModal
                 name=""
                 originalInput=""
-                setIsDownloadSpecModalOpen={function (condition: boolean): void {
+                setIsDownloadSpecModalOpen={function (): void {
                     throw new Error('Function not implemented.')
                 }}
-                setDownloadSpecModalDismissed={function (condition: boolean): void {
+                setDownloadSpecModalDismissed={function (): void {
                     throw new Error('Function not implemented.')
                 }}
                 {...props}
